requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
bokeh==2.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cloudpickle==1.5.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
cytoolz==0.10.1
dask==2.24.0
decorator==4.4.2
distributed==2.24.0
Django==3.0.3
django-extensions==2.2.9
djangorestframework==3.11.0
drf-yasg==1.17.1
fsspec==0.7.1
gast==0.3.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14
grpcio==1.28.1
h5py==2.10.0
HeapDict==1.0.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.1
jsonschema==3.2.0
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
locket==0.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
msgpack==1.0.0
networkx==2.4
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
partd==1.1.0
pathlib2==2.3.5
Pillow==7.0.0
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydotplus==2.0.2
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
seaborn==0.10.1
simplejson==3.17.0
six==1.14.0
sklearn
sortedcontainers==2.1.0
sqlparse==0.3.1
tblib==1.6.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0rc3
tensorflow-docs===0.0.058039b7770cd3bc90f12cf0a02c45968913c0153-
tensorflow-estimator==2.2.0
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.4
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
zict==2.0.0
